Fix multi enum column editor for enums with more than 8 entries
#rb samuele.rigamonti

[CL 35478510 by keith yerex in ue5-main branch]

+++ Engine/Plugins/Experimental/Chooser/Source/ChooserEditor/Private/MultiEnumColumnEditor.cpp
@@ -189,7 +189,7 @@ TSharedRef<SWidget> CreateMultiEnumColumnWidget(UChooserTable* Chooser, FChooser
 			{
 				if (MultiEnumColumn->RowValues.IsValidIndex(Row))
 				{
-					MultiEnumColumn->RowValues[Row].Value = static_cast<uint8>(Value);
+					MultiEnumColumn->RowValues[Row].Value = static_cast<uint32>(Value);
 				}
 			})
 			.EnumValue_Lambda([MultiEnumColumn, Row]()
